+++ index.js
@@ -1,1 +1,2 @@
 export { default as Path } from './src/Path.js';
+export { default as optimizeSvgPaths } from './src/optimizeSvgPaths.js';